requirements.txt
agate==1.9.1
alembic==1.12.1
annotated-types==0.7.0
anyio==4.10.0
asttokens==3.0.0
attrs==25.3.0
babel==2.17.0
certifi==2025.8.3
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.3
claude-monitor==3.1.0
click==8.2.1
colorama==0.4.6
comm==0.2.3
cryptography==41.0.7
daff==1.4.2
dbt-adapters==1.17.2
dbt-common==1.32.0
dbt-core==1.10.13
dbt-extractor==0.6.0
dbt-postgres==1.9.1
dbt-protos==1.0.382
dbt-semantic-interfaces==0.9.0
debugpy==1.8.16
decorator==5.2.1
deepdiff==8.6.1
docker==6.1.3
dotenv==0.9.9
exceptiongroup==1.3.0
executing==2.2.1
fastapi==0.115.12
fastmcp==2.2.0
git-filter-repo==2.47.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
hvac==2.1.0
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
isodate==0.6.1
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
leather==0.4.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mashumaro==3.14
matplotlib-inline==0.1.7
mcp==1.12.4
mdurl==0.1.2
more-itertools==10.8.0
msgpack==1.1.2
nest-asyncio==1.6.0
networkx==3.5
numpy==1.26.4
openapi-pydantic==0.5.1
orderly-set==5.5.0
packaging==25.0
pandas==2.1.3
parsedatetime==2.6
parso==0.8.5
pathspec==0.12.1
pdfminer.six==20250506
pdfplumber==0.11.7
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
pluggy==1.6.0
prompt_toolkit==3.0.52
protobuf==6.33.0
psutil==7.0.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
PyMuPDF==1.26.4
PyPDF2==3.0.1
pypdfium2==4.30.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.15.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.20
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
redis==5.0.1
referencing==0.36.2
requests==2.31.0
rich==14.1.0
rpds-py==0.27.1
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
snowplow-tracker==1.1.0
SQLAlchemy==2.0.23
sqlparse==0.5.3
sse-starlette==3.0.2
stack-data==0.6.3
starlette==0.46.2
structlog==25.4.0
text-unidecode==1.3
tornado==6.5.2
traitlets==5.14.3
typer==0.17.4
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.24.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
wheel==0.45.1
zipp==3.23.0

# ML/AI Libraries for homework week_04
torch>=2.6.0
torchvision>=0.21.0
scikit-learn>=1.3.2
matplotlib>=3.8.2
seaborn>=0.13.0
tqdm>=4.66.1
